pruned some duplicates
looks like I need a better way to detect or filter out stuff I don't want, mods with "~" in the name or "000","001", etc. 
Also commas in a mod name seem to break something (not very much, it still works, just thinks it's two mods instead of one, probably a filter error on "," somewhere.