ugly stuff to get mods linked up
parses the new dictionary text files to match up an ESP to a zipfile to a nexusID. Results are... not horrible. But if two mods have the same esp (like a patch), it will pick the first mod it sees.

Solution would be to write a proper nexusmods API using parser and downloader, but not today.